Initial Build
This is version 1.0 of code GameJam. There were many features added,
mostly HTML and CSS updates. A “sign-in” button and “sign-up” was added
to eventually onboard users to join a gaming community. The “You are
ready to P.A.C!” button entices users to sign-up and join the
community.  The main features include:

1. Images with text overlay
2. Quiz/Questionnaire
3. Accordion-style test results